tg_llm_bot/requirements.txt
python-telegram-bot==20.*
openai>=1.0.0